Code cleanup and lib Refactor
