+++ Seraphimdroid/app/build.gradle
@@ -39,7 +39,7 @@ dependencies {
     compile 'com.google.code.gson:gson:2.3.1'
 
     compile 'com.android.volley:volley:1.0.0'
-    compile(project(':lucenestudy')) {
+    compile(project(':lucenedroid')) {
         transitive = false
     }
 }
