+++ Seraphimdroid/lucenedroid/libs/lucene-core-5.3.0-mobile-2.jar
