+++ Seraphimdroid/lucenedroid/libs/lucene-highlighter-5.3.0-mobile-2.jar
