+++ Seraphimdroid/lucenedroid/libs/lucene-join-5.3.0-mobile-2.jar
