+++ Seraphimdroid/lucenedroid/libs/lucene-memory-5.3.0-mobile-2.jar
