+++ Seraphimdroid/lucenedroid/libs/lucene-misc-5.3.0-mobile-2.jar
