+++ Seraphimdroid/lucenedroid/libs/lucene-queries-5.3.0-mobile-2.jar
