+++ Seraphimdroid/lucenedroid/libs/lucene-queryparser-5.3.0-mobile-2.jar
